add CUSTOMER_HAS_VEHICLES validation

+++ src/main/java/br/com/fiap/postech/soat16/fase1/service/CustomerService.java
@@ -9,11 +9,13 @@ import jakarta.enterprise.context.ApplicationScoped;
 import br.com.fiap.postech.soat16.fase1.dto.pagination.PageableResponseDto;
 import br.com.fiap.postech.soat16.fase1.dto.request.CustomerRequestDto;
 import br.com.fiap.postech.soat16.fase1.dto.response.CustomerResponseDto;
+import br.com.fiap.postech.soat16.fase1.exception.CustomerHasVehiclesException;
 import br.com.fiap.postech.soat16.fase1.exception.CustomerNotFoundException;
 import br.com.fiap.postech.soat16.fase1.exception.DuplicateDocumentException;
 import br.com.fiap.postech.soat16.fase1.mapper.CustomerMapper;
 import br.com.fiap.postech.soat16.fase1.model.Document;
 import br.com.fiap.postech.soat16.fase1.repository.CustomerRepository;
+import br.com.fiap.postech.soat16.fase1.repository.VehicleRepository;
 
 import io.quarkus.hibernate.reactive.panache.common.WithSession;
 import io.quarkus.hibernate.reactive.panache.common.WithTransaction;
@@ -25,6 +27,7 @@ import lombok.RequiredArgsConstructor;
 public class CustomerService {
 
     private final CustomerRepository repository;
+    private final VehicleRepository vehicleRepository;
     private final CustomerMapper mapper;
 
     @WithSession
@@ -78,7 +81,13 @@ public class CustomerService {
 
     @WithTransaction
     public Uni<Void> delete(UUID id) {
-        return repository.deleteByCustomerId(id)
+        return vehicleRepository.existsByCustomerId(id)
+                .flatMap(hasVehicles -> {
+                    if (TRUE.equals(hasVehicles)) {
+                        return Uni.createFrom().<Long>failure(new CustomerHasVehiclesException());
+                    }
+                    return repository.deleteByCustomerId(id);
+                })
                 .flatMap(deleted -> {
                     if (deleted == 0) {
                         return Uni.createFrom().<Void>failure(new CustomerNotFoundException());
